Track feedback usage on the prompt-submit hook, not the stop hook
Move recordActivity from the stop hook to the auto (UserPromptSubmit) hook, so
active usage is counted once per genuine user prompt (after the advisory-injected
guard). The stop hook now only checks eligibility and shows the popup.
