+++ src/cli/commands/stop-signal-capture.test.ts
@@ -25,7 +25,6 @@ import type { SelectFn } from '../../decision-session/DecisionSession.js';
 import * as TtySelectFnModule from '../../decision-session/TtySelectFn.js';
 import { setConfig } from '../../store/config.js';
 import { readSignals } from '../../store/feedback-signals.js';
-import { readCadence } from '../../store/feedback-cadence.js';
 
 const CWD = '/test/project';
 
@@ -59,42 +58,20 @@ let store: Store;
 beforeEach(async () => { store = await openStore(':memory:'); });
 afterEach(() => { closeStore(store); vi.restoreAllMocks(); });
 
-describe('usage recording', () => {
-  it('records activity on a normal turn (even with no advisory)', async () => {
-    await runStop(makePayload(), store);
-    expect(readCadence(store).lastActivityAt).not.toBeNull();
-  });
-
-  it('does not record activity on the loop guard', async () => {
-    await runStop(makePayload({ stop_hook_active: true }), store);
-    expect(readCadence(store).lastActivityAt).toBeNull();
-  });
-
-  it('accumulates usage globally across turns within the idle cap', async () => {
-    const nowSpy = vi.spyOn(Date, 'now');
-    nowSpy.mockReturnValue(1_000_000);
-    await runStop(makePayload({ cwd: '/proj-a' }), store);
-    nowSpy.mockReturnValue(1_000_000 + 60_000);
-    await runStop(makePayload({ cwd: '/proj-b' }), store); // different project → same global counter
-    expect(readCadence(store).activeMs).toBe(60_000);
-  });
-
-  it('records activity even when the advisory frequency is off', async () => {
+describe('advisory not shown → no fire recorded', () => {
+  it('does not record a fire when the advisory frequency is off', async () => {
     setConfig(store, 'advisory_frequency', 'off');
     insertAdvisory(store);
     const result = await runStop(makePayload(), store, mockSelect(SKIP_NOW));
     expect(result.outcome).toBe('skipped');
-    expect(readCadence(store).lastActivityAt).not.toBeNull();
-    // Freq-gated advisory is never shown → no fire recorded.
     expect(readSignals(store, CWD).advisoryFireTs).toHaveLength(0);
   });
 
-  it('records activity but no fire when there is no TTY', async () => {
+  it('does not record a fire when there is no TTY', async () => {
     insertAdvisory(store);
     vi.spyOn(TtySelectFnModule, 'createTtySelectFn').mockReturnValue(null);
-    const result = await runStop(makePayload(), store); // no selectFn → resolves TTY
+    const result = await runStop(makePayload(), store);
     expect(result.outcome).toBe('no_tty');
-    expect(readCadence(store).lastActivityAt).not.toBeNull();
     expect(readSignals(store, CWD).advisoryFireTs).toHaveLength(0);
   });
 });
